fix: Handling arrays has been fixed
Arrays were simply not being handled correctly. They detect as objects but
I never checked specifically for them, so it would ugily list them out
with the indexes as attributes. This would cause major issues in some
applications, so I added logic to handle if both merge objects are arrays
or if one or the other was an array.